Add app/reunion-client.py and app/reunion-server.py These implement HTTP and TCP server/clients for running REUNION over a network.
There are some preliminary Tor integration doodles in the server.

The client outputs its results as newline-terminated JSON tuples,
this should be documented.

The time stuff (delays etc) could use some more love; sub-second defaults are probably not great for real-world usage.

Both applications are also currently missing anything in the realm of epoch handling.
The client doesn't really handle the server being restarted; it should resend stuff that should be on the server but isn't, at least if the epochs are the same. It should probably also alert about that, since it's a potential attack (that warning would be a lot more useful if the server persisted to disk).

Another thing that should be debugged if we answer too many T1/T2/T3s.

In short: we've a while to go, but we are getting there.