Add setup.py and update documentation
This changeset adds a basic setup.py. The setup.py is derived from the
related Vula project by two of the same authors. The license is
identical and the setup.py has been slightly refactored. The setup.py
extracts the version from the updated pyproject.toml file.

Important use cases include:

  python3 setup.py install
  python3 setup.py test
  python3 setup.py version
  python3 setup.py --command-packages=stdeb.command sdist_dsc bdist_deb

The final command above produces a binary Debian package which is
suitable for installation on Debian systems.

The two programs in app have been moved into the module and added as
entrypoints in the setup.py. After this commit three commands will be
installed by the setup.py and also included in the Debian package:

  reunion
  reunion-server
  reunion-on-an-ethernet