Rename module; update module and metadata for pypi
This commit updates the Python package from reunion to rendez due to the
namespace available on pypi. Unfortunately, many of the relevant package
names such as `reunion`, or `rendezovus` are taken on pypi. Fortunately,
the root of rendezvous is available! This allows the `reunion` module to
be imported from the `rendez` package as `from rendez.vous import
reunion` or `import rendez.vous.reunion` or `import rendez.vous.reunion
as reunion`.

- Tests use the new import path
- The `reunion` module use the new import path
- The Debian packaging has been updated to reflect the new Python
  package name. It produces a new Debian package name:
  `python3-rendez_*_all.deb`
- The `.gitignore` has been updated to ignore a commonly used venv name
- The `pyproject.toml` has been updated to reflect the new package name
- The `Makefile` has been updated to be more generic, to remove
  `__pycache__` more thoroughly, it has a new target for checking the
  sdist before uploading to pypi (`pypi-check`) and a new target to
  upload to pypi (`pypi-upload`). This target checks the source package
  with `twine` before uploading